Port it to Python 2.
Python 2 will be discontinued in 2020, but Flask and a lot of its
plugins run only with this version.

The code is commented so it's easy to 2to3 again. This should help the
consistency of the project too, since it won't have parts of it running
on Python 2 and parts of it running on Python 3.